Relocates functions
