--- internal/service/s3/bucket_lifecycle_configuration.go
@@ -1,147 +1,0 @@
-// Copyright (c) HashiCorp, Inc.
-// SPDX-License-Identifier: MPL-2.0
-
-package s3
-
-import (
-	"context"
-	"slices"
-	"strconv"
-	"time"
-
-	"github.com/aws/aws-sdk-go-v2/aws"
-	"github.com/aws/aws-sdk-go-v2/service/s3"
-	"github.com/aws/aws-sdk-go-v2/service/s3/types"
-	"github.com/hashicorp/aws-sdk-go-base/v2/tfawserr"
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/retry"
-	"github.com/hashicorp/terraform-provider-aws/internal/tfresource"
-	"github.com/hashicorp/terraform-provider-aws/names"
-)
-
-func findBucketLifecycleConfiguration(ctx context.Context, conn *s3.Client, bucket, expectedBucketOwner string) (*s3.GetBucketLifecycleConfigurationOutput, error) {
-	input := &s3.GetBucketLifecycleConfigurationInput{
-		Bucket: aws.String(bucket),
-	}
-	if expectedBucketOwner != "" {
-		input.ExpectedBucketOwner = aws.String(expectedBucketOwner)
-	}
-
-	output, err := conn.GetBucketLifecycleConfiguration(ctx, input)
-
-	if tfawserr.ErrCodeEquals(err, errCodeNoSuchBucket, errCodeNoSuchLifecycleConfiguration) {
-		return nil, &retry.NotFoundError{
-			LastError:   err,
-			LastRequest: input,
-		}
-	}
-
-	if err != nil {
-		return nil, err
-	}
-
-	if output == nil || len(output.Rules) == 0 {
-		return nil, tfresource.NewEmptyResultError(input)
-	}
-
-	return output, nil
-}
-
-func lifecycleRulesEqual(rules1, rules2 []types.LifecycleRule) bool {
-	if len(rules1) != len(rules2) {
-		return false
-	}
-
-	for _, rule1 := range rules1 {
-		// We consider 2 LifecycleRules equal if their IDs and Statuses are equal.
-		if !slices.ContainsFunc(rules2, func(rule2 types.LifecycleRule) bool {
-			return aws.ToString(rule1.ID) == aws.ToString(rule2.ID) && rule1.Status == rule2.Status
-		}) {
-			return false
-		}
-	}
-
-	return true
-}
-
-func statusLifecycleRulesEquals(ctx context.Context, conn *s3.Client, bucket, expectedBucketOwner string, rules []types.LifecycleRule) retry.StateRefreshFunc {
-	return func() (interface{}, string, error) {
-		output, err := findBucketLifecycleConfiguration(ctx, conn, bucket, expectedBucketOwner)
-
-		if tfresource.NotFound(err) {
-			return nil, "", nil
-		}
-
-		if err != nil {
-			return nil, "", err
-		}
-
-		return output, strconv.FormatBool(lifecycleRulesEqual(output.Rules, rules)), nil
-	}
-}
-
-func waitLifecycleRulesEquals(ctx context.Context, conn *s3.Client, bucket, expectedBucketOwner string, rules []types.LifecycleRule, timeout time.Duration) ([]types.LifecycleRule, error) { //nolint:unparam
-	stateConf := &retry.StateChangeConf{
-		Target:                    []string{strconv.FormatBool(true)},
-		Refresh:                   statusLifecycleRulesEquals(ctx, conn, bucket, expectedBucketOwner, rules),
-		Timeout:                   timeout,
-		MinTimeout:                10 * time.Second,
-		ContinuousTargetOccurence: 3,
-		NotFoundChecks:            20,
-	}
-
-	outputRaw, err := stateConf.WaitForStateContext(ctx)
-
-	if output, ok := outputRaw.([]types.LifecycleRule); ok {
-		return output, err
-	}
-
-	return nil, err
-}
-
-const (
-	lifecycleRuleStatusDisabled = "Disabled"
-	lifecycleRuleStatusEnabled  = "Enabled"
-)
-
-func lifecycleRuleStatus_Values() []string {
-	return []string{
-		lifecycleRuleStatusDisabled,
-		lifecycleRuleStatusEnabled,
-	}
-}
-
-func expandTag(tfMap map[string]interface{}) *types.Tag {
-	if len(tfMap) == 0 {
-		return nil
-	}
-
-	apiObject := &types.Tag{}
-
-	if v, ok := tfMap[names.AttrKey].(string); ok {
-		apiObject.Key = aws.String(v)
-	}
-
-	if v, ok := tfMap[names.AttrValue].(string); ok {
-		apiObject.Value = aws.String(v)
-	}
-
-	return apiObject
-}
-
-func flattenTag(apiObject *types.Tag) []interface{} {
-	if apiObject == nil {
-		return nil
-	}
-
-	tfMap := make(map[string]interface{})
-
-	if v := apiObject.Key; v != nil {
-		tfMap[names.AttrKey] = aws.ToString(v)
-	}
-
-	if v := apiObject.Value; v != nil {
-		tfMap[names.AttrValue] = aws.ToString(v)
-	}
-
-	return []interface{}{tfMap}
-}
