+++ internal/service/s3/bucket_lifecycle_configuration_fw.go
@@ -6,6 +6,8 @@ package s3
 import (
 	"context"
 	"fmt"
+	"slices"
+	"strconv"
 	"time"
 
 	"github.com/aws/aws-sdk-go-v2/aws"
@@ -601,6 +603,98 @@ func (r *resourceBucketLifecycleConfiguration) UpgradeState(ctx context.Context)
 			PriorSchema:   &schemaV0,
 			StateUpgrader: upgradeBucketLifeCycleConfigurationResourceStateV0toV1,
 		},
+	}
+}
+
+func findBucketLifecycleConfiguration(ctx context.Context, conn *s3.Client, bucket, expectedBucketOwner string) (*s3.GetBucketLifecycleConfigurationOutput, error) {
+	input := &s3.GetBucketLifecycleConfigurationInput{
+		Bucket: aws.String(bucket),
+	}
+	if expectedBucketOwner != "" {
+		input.ExpectedBucketOwner = aws.String(expectedBucketOwner)
+	}
+
+	output, err := conn.GetBucketLifecycleConfiguration(ctx, input)
+
+	if tfawserr.ErrCodeEquals(err, errCodeNoSuchBucket, errCodeNoSuchLifecycleConfiguration) {
+		return nil, &retry.NotFoundError{
+			LastError:   err,
+			LastRequest: input,
+		}
+	}
+
+	if err != nil {
+		return nil, err
+	}
+
+	if output == nil || len(output.Rules) == 0 {
+		return nil, tfresource.NewEmptyResultError(input)
+	}
+
+	return output, nil
+}
+
+func lifecycleRulesEqual(rules1, rules2 []awstypes.LifecycleRule) bool {
+	if len(rules1) != len(rules2) {
+		return false
+	}
+
+	for _, rule1 := range rules1 {
+		// We consider 2 LifecycleRules equal if their IDs and Statuses are equal.
+		if !slices.ContainsFunc(rules2, func(rule2 awstypes.LifecycleRule) bool {
+			return aws.ToString(rule1.ID) == aws.ToString(rule2.ID) && rule1.Status == rule2.Status
+		}) {
+			return false
+		}
+	}
+
+	return true
+}
+
+func statusLifecycleRulesEquals(ctx context.Context, conn *s3.Client, bucket, expectedBucketOwner string, rules []awstypes.LifecycleRule) retry.StateRefreshFunc {
+	return func() (interface{}, string, error) {
+		output, err := findBucketLifecycleConfiguration(ctx, conn, bucket, expectedBucketOwner)
+
+		if tfresource.NotFound(err) {
+			return nil, "", nil
+		}
+
+		if err != nil {
+			return nil, "", err
+		}
+
+		return output, strconv.FormatBool(lifecycleRulesEqual(output.Rules, rules)), nil
+	}
+}
+
+func waitLifecycleRulesEquals(ctx context.Context, conn *s3.Client, bucket, expectedBucketOwner string, rules []awstypes.LifecycleRule, timeout time.Duration) ([]awstypes.LifecycleRule, error) { //nolint:unparam
+	stateConf := &retry.StateChangeConf{
+		Target:                    []string{strconv.FormatBool(true)},
+		Refresh:                   statusLifecycleRulesEquals(ctx, conn, bucket, expectedBucketOwner, rules),
+		Timeout:                   timeout,
+		MinTimeout:                10 * time.Second,
+		ContinuousTargetOccurence: 3,
+		NotFoundChecks:            20,
+	}
+
+	outputRaw, err := stateConf.WaitForStateContext(ctx)
+
+	if output, ok := outputRaw.([]awstypes.LifecycleRule); ok {
+		return output, err
+	}
+
+	return nil, err
+}
+
+const (
+	lifecycleRuleStatusDisabled = "Disabled"
+	lifecycleRuleStatusEnabled  = "Enabled"
+)
+
+func lifecycleRuleStatus_Values() []string {
+	return []string{
+		lifecycleRuleStatusDisabled,
+		lifecycleRuleStatusEnabled,
 	}
 }
